chore(): Rename path
